+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/texture/V2/TextureService.java
@@ -11,33 +11,31 @@ public class TextureService {
     public static String PALETTE_PATH = "assets/forgero/templates/materials/";
     public static String TEMPLATE_PATH = "assets/forgero/templates/textures/";
     private final TextureLoader loader;
+    private final Map<String, String> paletteRemap;
     private Map<String, Palette> paletteCache;
     private Map<String, TemplateTexture> templateCache;
-    private Map<String, Palette> previousPaletteCache;
-    private Map<String, TemplateTexture> previousTemplateCache;
 
-    public TextureService(FileLoader loader) {
+    public TextureService(FileLoader loader, Map<String, String> paletteRemap) {
         this.paletteCache = new HashMap<>();
+        this.paletteRemap = paletteRemap;
         this.templateCache = new HashMap<>();
-        this.previousPaletteCache = new HashMap<>();
-        this.previousTemplateCache = new HashMap<>();
         this.loader = new TextureLoader(loader);
     }
 
     public Optional<Palette> getPalette(String name) {
-        
+        var remapped = Optional.ofNullable(paletteRemap.get(name)).flatMap(this::getPalette);
+        if (remapped.isPresent()) {
+            return remapped;
+        }
         if (paletteCache.containsKey(name)) {
             return Optional.of(paletteCache.get(name));
         }
-        var template = loader.load(PALETTE_PATH + name).or(() -> Optional.ofNullable(previousTemplateCache.get(name).getImage()));
+        var template = loader.load(PALETTE_PATH + name);
         template.ifPresent(palette -> paletteCache.put(name, new Palette(palette)));
         return template.map(Palette::new);
     }
 
     public void clear() {
-        previousPaletteCache = paletteCache;
-        previousTemplateCache = templateCache;
-
         templateCache = new ConcurrentHashMap<>();
         paletteCache = new ConcurrentHashMap<>();
     }