+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/type/Type.java
@@ -19,13 +19,13 @@ public interface Type extends Matchable {
 
     Type PART = new SimpleType("PART", Optional.empty(), new TypeMatcher());
     Type BINDING = new SimpleType("BINDING", Optional.of(PART), new TypeMatcher());
+    Type SWORD_GUARD = new SimpleType("SWORD_GUARD", Optional.of(PART), new TypeMatcher());
     Type TOOL_PART_HEAD = new SimpleType("TOOL_PART_HEAD", Optional.of(PART), new TypeMatcher());
     Type AXE_HEAD = new SimpleType("AXE_HEAD", Optional.of(TOOL_PART_HEAD), new TypeMatcher());
     Type HANDLE = new SimpleType("HANDLE", Optional.empty(), new TypeMatcher());
 
     Type BLADE = new SimpleType("BLADE", Optional.empty(), new TypeMatcher());
     Type SWORD_BLADE = new SimpleType("SWORD_BLADE", Optional.of(BLADE), new TypeMatcher());
-
 
     Type SCHEMATIC = new SimpleType("SCHEMATIC", Optional.empty(), new SchematicMatcher());