+++ forgero-fabric/src/main/java/com/sigmundgranaas/forgero/fabric/mixins/PlayerServerInteractionManagerMixin.java
@@ -1,23 +1,11 @@
 package com.sigmundgranaas.forgero.fabric.mixins;
 
-import com.sigmundgranaas.forgero.core.property.ActivePropertyType;
-import com.sigmundgranaas.forgero.core.property.Property;
-import com.sigmundgranaas.forgero.core.property.TargetTypes;
-import com.sigmundgranaas.forgero.core.property.active.VeinBreaking;
-import com.sigmundgranaas.forgero.core.property.attribute.SingleTarget;
-import com.sigmundgranaas.forgero.core.state.State;
-import com.sigmundgranaas.forgero.minecraft.common.item.StateItem;
-import com.sigmundgranaas.forgero.minecraft.common.property.handler.PatternBreaking;
-import com.sigmundgranaas.forgero.minecraft.common.toolhandler.BlockBreakingHandler;
-import com.sigmundgranaas.forgero.minecraft.common.toolhandler.PatternBreakingStrategy;
-import com.sigmundgranaas.forgero.minecraft.common.toolhandler.SoulHandler;
-import com.sigmundgranaas.forgero.minecraft.common.toolhandler.VeinMiningStrategy;
-import net.minecraft.block.BlockState;
+import com.sigmundgranaas.forgero.minecraft.common.toolhandler.PropertyHelper;
+import com.sigmundgranaas.forgero.minecraft.common.toolhandler.ToolBlockHandler;
 import net.minecraft.network.packet.c2s.play.PlayerActionC2SPacket;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.network.ServerPlayerInteractionManager;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.util.Pair;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import org.spongepowered.asm.mixin.Final;
@@ -26,9 +14,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-import java.util.Collections;
-import java.util.List;
 
 @Mixin(ServerPlayerInteractionManager.class)
 public abstract class PlayerServerInteractionManagerMixin {
@@ -43,25 +28,8 @@ public abstract class PlayerServerInteractionManagerMixin {
 
     @Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/server/network/ServerPlayerInteractionManager;finishMining(Lnet/minecraft/util/math/BlockPos;ILjava/lang/String;)V"), method = "processBlockBreakingAction")
     public void processBlockBreakingAction(BlockPos pos, PlayerActionC2SPacket.Action action, Direction direction, int worldHeight, int sequence, CallbackInfo ci) {
-        if (player.getMainHandStack().getItem() instanceof StateItem stateItem) {
-            State toolState = stateItem.dynamicState(player.getMainHandStack());
-            var activeProperties = Property.stream(toolState.applyProperty(new SingleTarget(TargetTypes.BLOCK, Collections.emptySet()))).getActiveProperties().toList();
-            SoulHandler.of(player.getMainHandStack()).ifPresent(handler -> handler.processBlockBreak(world.getBlockState(pos), pos, world, player));
-            if (!activeProperties.isEmpty()) {
-                List<Pair<BlockState, BlockPos>> availableBlocks;
-                if (activeProperties.get(0).getActiveType() == ActivePropertyType.BLOCK_BREAKING_PATTERN) {
-                    availableBlocks = new BlockBreakingHandler(new PatternBreakingStrategy((PatternBreaking) activeProperties.get(0))).getAvailableBlocks(world, pos, player);
-                } else {
-                    availableBlocks = new BlockBreakingHandler(new VeinMiningStrategy((VeinBreaking) activeProperties.get(0))).getAvailableBlocks(world, pos, player);
-                }
-                for (var block : availableBlocks) {
-                    if (!block.getRight().equals(pos)) {
-                        SoulHandler.of(player.getMainHandStack()).ifPresent(handler -> handler.processBlockBreak(block.getLeft(), block.getRight(), world, player));
-                        this.finishMining(block.getRight(), sequence, "destroyed");
-
-                    }
-                }
-            }
-        }
+        PropertyHelper.ofPlayerHands(player)
+                .flatMap(container -> ToolBlockHandler.of(container, world, pos, player))
+                .ifPresent(handler -> handler.handleExceptOrigin(info -> this.finishMining(info.pos(), sequence, "destroyed")));
     }
 }