+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/tooltip/AttributeWriter.java
@@ -86,13 +86,14 @@ public class AttributeWriter implements Writer {
             case RARITY -> intAttribute(RARITY, tooltip);
             case DURABILITY -> intAttribute(DURABILITY, tooltip);
             case MINING_LEVEL -> intAttribute(MINING_LEVEL, tooltip);
+            case ARMOR -> floatAttribute(ARMOR, tooltip);
         }
     }
 
     private void intAttribute(AttributeType type, List<Text> tooltip) {
         int result = (int) helper.attribute(type);
         if (result != 0) {
-            MutableText miningLevel = Text.literal("  ").append(Text.translatable(Writer.toTranslationKey(type.toString()))).append(": ").formatted(Formatting.GRAY);
+            MutableText miningLevel = Text.literal("  ").append(Text.translatable(Writer.toTranslationKey(type.toString().toLowerCase()))).append(": ").formatted(Formatting.GRAY);
             miningLevel.append(Text.literal(String.format("%s", result)).formatted(Formatting.WHITE));
             tooltip.add(miningLevel);
         }
@@ -105,7 +106,7 @@ public class AttributeWriter implements Writer {
         }
         result = roundFloat(result);
         if (result != 0f) {
-            MutableText miningLevel = Text.literal("  ").append(Text.translatable(Writer.toTranslationKey(type.toString()))).append(": ").formatted(Formatting.GRAY);
+            MutableText miningLevel = Text.literal("  ").append(Text.translatable(Writer.toTranslationKey(type.toString().toLowerCase()))).append(": ").formatted(Formatting.GRAY);
             miningLevel.append(Text.literal(String.format("%s", result)).formatted(Formatting.WHITE));
             tooltip.add(miningLevel);
         }